refactor: moved constant to a utils file to be easily reused on test file
